# api/openapi/public/openapi.yaml
openapi: 3.0.0
info:
  title: E-Commerce Platform Public API
  version: 0.1.0
  description: Initial version of REST API surface.
servers:
  - url: http://127.0.0.1:8080
    description: Local gateway
...
security:
  - bearerAuth: []
paths:
  /v1/auth/register:
    post:
      tags: [Auth]
      operationId: registerUser
      summary: Register a new user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokensResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/auth/register-admin:
    post:
      tags: [Auth]
      operationId: registerAdmin
      summary: Register a new admin user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Admin user registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokensResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/auth/login:
    post:
      tags: [Auth]
      operationId: loginUser
      summary: Authenticate user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokensResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/auth/refresh:
    post:
      tags: [Auth]
      operationId: refreshToken
      summary: Refresh access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Tokens refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokensResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/profile/me:
    get:
      tags: [Profile]
      operationId: getMyProfile
      summary: Get current user profile
      responses:
        '200':
          description: Profile details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'
    patch:
      tags: [Profile]
      operationId: updateMyProfile
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated caller lacks required permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        correlationId:
          type: string
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          maxLength: 254
          x-oapi-codegen-extra-tags:
            validate: "required,email,max=254"
        password:
          type: string
          minLength: 8
          maxLength: 72
          x-oapi-codegen-extra-tags:
            validate: "required,min=8,max=72"
        displayName:
          type: string
          minLength: 1
          maxLength: 100
          x-oapi-codegen-extra-tags:
            validate: "omitempty,min=1,max=100"
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          maxLength: 254
          x-oapi-codegen-extra-tags:
            validate: "required,email,max=254"
        password:
          type: string
          minLength: 1
          maxLength: 72
          x-oapi-codegen-extra-tags:
            validate: "required,min=1,max=72"
    RefreshTokenRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
          minLength: 1
          x-oapi-codegen-extra-tags:
            validate: "required"
    AuthTokensResponse:
      type: object
      required: [accessToken, refreshToken]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
    UserProfile:
      type: object
      required: [userId, email, role, status]
      properties:
        userId:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [user, admin]
        status:
          type: string
          enum: [active, disabled]
        displayName:
          type: string
    UpdateProfileRequest:
      type: object
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 100
          x-oapi-codegen-extra-tags:
            validate: "omitempty,min=1,max=100"
...